Refactor activity package
