+++ telegrambots-extensions/src/main/java/org/telegram/telegrambots/extensions/bots/commandbot/activity/DefaultUserActivityHandler.java
@@ -1,4 +1,4 @@
-package org.telegram.telegrambots.extensions.bots.commandbot.commands.activity;
+package org.telegram.telegrambots.extensions.bots.commandbot.activity;
 
 import java.util.HashMap;
 import java.util.Map;
